broke nlocal into separate files
